fix: chip to pass input event
